add vlc shims
