+++ Gemfile
@@ -17,6 +17,7 @@ gem 'thin'
 if RUBY_PLATFORM.downcase.include?("darwin")
   gem 'rb-appscript'
 end
+gem 'vlcrc'
 
 group :test do
   gem 'rack-test'
