+++ app/boot.rb
@@ -21,6 +21,8 @@ require 'models/album'
 require 'models/history'
 require 'models/office'
 require 'models/player'
+require 'models/player/itunes'
+require 'models/player/vlc'
 require 'models/queue'
 require 'models/song'
 require 'models/user'
