+++ app/models/queue.rb
