+++ app/play.rb
@@ -15,7 +15,8 @@ module Play
       :pusher_app_id => yaml['pusher_app_id'],
       :pusher_key    => yaml['pusher_key'],
       :pusher_secret => yaml['pusher_secret'],
-      :auth_token    => yaml['auth_token']
+      :auth_token    => yaml['auth_token'],
+      :player        => yaml['player']
   end
 
 private
